+++ src/k8s-config/HISTORY.rst
@@ -5,4 +5,5 @@ Release History
 
 0.1.0
 ++++++
-* Initial release.
+
+* Release k8s-config with flux configuration support